+++ packages/app/src/Pages/Profile/ProfileTabSelectors.tsx
@@ -1,10 +1,10 @@
 import { FormattedMessage } from "react-intl";
 
 import Icon from "@/Components/Icons/Icon";
-import { Tab } from "@/Components/Tabs/Tabs";
+import { Tab } from "@/Components/TabSelectors/TabSelectors";
 import { ProfileTabType } from "@/Pages/Profile/ProfileTabType";
 
-const ProfileTab = {
+const ProfileTabSelectors = {
   Notes: {
     text: (
       <>
@@ -88,4 +88,4 @@ const ProfileTab = {
   },
 } as { [key: string]: Tab };
 
-export default ProfileTab;
+export default ProfileTabSelectors;